% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMphyClusInterface.R
\docType{data}
\name{betweenClusTransProbs}
\alias{betweenClusTransProbs}
\title{Between-cluster transition probability matrices}
\format{A list of length 10. Each list element is itself a list, but with
length 3, corresponding to the number of rate variation categories. Each
element of the nested list is a 4x4 matrix giving transition probabilities
between the different nucleotide states.}
\source{
\url{http://mbe.oxfordjournals.org/content/18/6/897.short}
}
\usage{
betweenClusTransProbs
}
\description{
These matrices are meant to be used in the vignette example. They were
obtained with \code{outputTransMatList}. They represent
transition probabilities along branches in the between-cluster (supporting)
phylogeny. Among-sites rate variation follows a discrete gamma distribution
with 3 categories, and with parameter 0.7589 (Posada et al. 2001). We assumed
that branch lengths follow the log-normal distribution with means between 0.015
and 0.045 (10 equally spaced mean assumptions).
}
\keyword{datasets}

