% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMphyClusInterface.R
\name{clusIndLogPrior}
\alias{clusIndLogPrior}
\title{Log-prior probability for a cluster assignment indices vector.}
\usage{
clusIndLogPrior(clusInd, alpha, k)
}
\arguments{
\item{clusInd}{vector of cluster assingment indices (should not contain any gaps)}

\item{alpha}{positive number, concentration parameter in the Dirichlet distribution}

\item{k}{positive integer giving the number of clusters; if not provided by the user, assumed to be equal to max(clusInd)}
}
\value{
A negative numeric value.
}
\description{
This function outputs the log-prior probability for a given cluster assignment index vector, concentration parameter, and number of clusters. It is mostly used for verification purposes.
}
\examples{
\dontrun{
clusIndLogPrior(clusInd = c(1,2,2,3,4,4,4), alpha = 1)
}
}
