% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMphyClusInterface.R
\name{logLikFromClusInd}
\alias{logLikFromClusInd}
\title{Phylogenetic log-likelihood}
\usage{
logLikFromClusInd(phylogeny, betweenTransMatList, withinTransMatList, clusInd,
  limProbs, numLikThreads, alignment, basicChecks = TRUE)
}
\arguments{
\item{phylogeny}{phylo object}

\item{betweenTransMatList}{list of matrices giving transition probabilities in the supporting phylogeny by rate variation category}

\item{withinTransMatList}{list of matrices giving transition probabilities in the within-cluster phylogenies by rate variation category}

\item{clusInd}{vector of cluster assingment indices (should not contain any gaps)}

\item{limProbs}{a vector giving limiting probabilities for all states in alignment}

\item{numLikThreads}{positive integer giving the number of threads used}

\item{alignment}{a character matrix with rows corresponding to sequences and columns to sites}

\item{basicChecks}{boolean indicating whether basic checks should be performed to ensure the call does not result in a segmentation fault}
}
\value{
A negative value.
}
\description{
This function gives the log-likelihood for a given alignment, conditional on a phylogeny and a cluster assignment index vector. It's very slow and should only be used for testing purposes.
}
\details{
This function is an alternative to logLikFromSplitPhylo. Just like that function, it relies on compiled C++ code and can cause segmentation faults if the supplied phylo objects are misformed. Basic checks are performed by default. The function will be faster if they are turned off.
}
\examples{
\dontrun{
INPUT_EXAMPLE()
}
}
