% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMphyClusInterface.R
\name{logLikFromSplitPhylo}
\alias{logLikFromSplitPhylo}
\title{Phylogenetic log-likelihood}
\usage{
logLikFromSplitPhylo(clusterPhylos, betweenPhylo, betweenTransMatList,
  withinTransMatList, clusInd, alignment, limProbs, numLikThreads = 1,
  basicChecks = TRUE)
}
\arguments{
\item{clusterPhylos}{list of phylo objects or NULL (indicating a singleton) with names matching tip labels in betweenPhylo}

\item{betweenPhylo}{phylo object with tip labels matching list names in clusterPhylos}

\item{betweenTransMatList}{list of matrices giving transition probabilities in the supporting phylogeny by rate variation category}

\item{withinTransMatList}{list of matrices giving transition probabilities in the within-cluster phylogenies by rate variation category}

\item{clusInd}{vector of cluster assingment indices (should not contain any gaps)}

\item{alignment}{a character matrix with rows corresponding to sequences and columns to sites}

\item{limProbs}{a vector giving limiting probabilities for all states in alignment}

\item{numLikThreads}{positive integer giving the number of threads used}

\item{basicChecks}{boolean indicating whether basic checks should be performed to ensure the call does not result in a segmentation fault}
}
\value{
A negative value.
}
\description{
This function gives the log-likelihood for a given alignment, conditional on a phylogeny split in between- and within- cluster components.
}
\details{
This function relies on compiled C++ code and can cause segmentation faults if the supplied phylo objects are misformed. Basic checks are performed by default. The function will be faster if they are turned off.
}
\examples{
\dontrun{
INPUT_EXAMPLE()
}

}

