% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMphyClusInterface.R
\name{outputTransMatList}
\alias{outputTransMatList}
\title{Obtaining transition probability matrices}
\usage{
outputTransMatList(QmatScaled, numGammaCat, gammaShape, numReplicates = 10^5,
  distRanFun, ...)
}
\arguments{
\item{QmatScaled}{a matrix indicating transition rates among all potential states; should be scaled in such a way that each row sums to 0 and -diag(QmatScaled)*(chain limiting probabilities) = 1.}

\item{numGammaCat}{positive integer giving the number of rate variation categories}

\item{gammaShape}{positive number giving the value of the discrete gamma shape parameter}

\item{numReplicates}{positive integer giving the number of random values used in obtaining the estimate}

\item{distRanFun}{function for simulating random values, must have an argument called 'n', standing for the number of values to simulate, e.g. rnorm,rexp}

\item{...}{additional parameters for distRanFun}
}
\value{
A list with length equal to numGammaCat. Each element is a matrix with dimensions equal to those of QmatScaled.
}
\description{
This function produces a Monte Carlo estimate of transition probability matrices under the assumption of discrete gamma among-sites rate variation.
}
\details{
This function produces estimates of the transition probability matrices along branches in the between- or within- cluster phylogenies. It is called by DMphyClus when betweenClusTransMatList and withinClusTransMatList are not provided. The estimation can take a long time, so we recommend doing it only once with this function, saving the output, and re-using it whenever necessary.
}
\examples{
\dontrun{
INPUT_AN_EXAMPLE()
}
}

