% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMphyClusInterface.R
\docType{data}
\name{QmatrixPosada2001}
\alias{QmatrixPosada2001}
\title{Nucleotide substitution rate matrix}
\format{A matrix with 4 rows and 4 columns.}
\source{
\url{http://mbe.oxfordjournals.org/content/18/6/897.short}
}
\description{
This matrix, from Posada et al. 2001, gives nucleotide substitution rates
in a standard HIV-1 alignment.
}
